move aflowprototype to third_party module
