+++ src/simmate/configuration/django/settings.py
@@ -164,7 +164,6 @@ INSTALLED_APPS = [
     "simmate.website.third_parties.apps.ThirdPartyConfig",
     "simmate.website.workflows.apps.WorkflowsConfig",
     "simmate.website.workflow_execution.apps.WorkflowExecutionConfig",
-    "simmate.website.prototypes.apps.PrototypesConfig",
     #
     # These are built-in django apps that we use for extra features
     "django.contrib.admin",
@@ -196,9 +195,7 @@ INSTALLED_APPS = [
     # need to be installed apps while some also request different setups.
     #   django-ratelimit
     #   dj-stripe
-    #   django-allauth
     #   django-unicorn
-    #   django-REST-framework
     #   django-graphene (+ GraphQL)
     #   django-redis
     #
