--- src/simmate/database/prototypes/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from simmate.utilities import get_doc_from_readme
-
-__doc__ = get_doc_from_readme(__file__)
-
-from .aflow import AflowPrototype
