+++ src/simmate/database/third_parties/aflow/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from .structures import AflowStructure
+from .prototypes import AflowPrototype
