+++ src/simmate/database/third_parties/aflow/prototypes.py
@@ -11,22 +11,28 @@ class AflowPrototype(Structure):
     as a table of structure-types. All of the entries in this table come
     from the AFLOW Encyclopedia of Crystallographic Prototypes:
       http://www.aflowlib.org/prototype-encyclopedia/
+
+    Note, while AFLOW licensing does not allow redistribution, all data in this
+    table is grabbed directly from `pymatgen`, where the data is under a
+    different license. This data does NOT include the entire encyclopedia and
+    is therefore out of date. For more info, see
+    [pymatgen issue #1446](https://github.com/materialsproject/pymatgen/issues/1446)
     """
 
     class Meta:
-        app_label = "prototypes"
+        app_label = "third_parties"
 
     base_info = [
         "structure_string",
         "mineral_name",
         "aflow_id",
         "pearson_symbol",
-        "strukturbericht",
+        "strukturbericht_symbol",
         "nsites_wyckoff",
     ]
     source = "AFLOW Encyclopedia of Crystallographic Prototypes"
     source_doi = "https://doi.org/10.1016/j.commatsci.2017.01.017"
-    # remote_archive_link = "TODO"  # also this will be private due to AFLOW licensing
+    remote_archive_link = "https://archives.simmate.org/AflowPrototype-2022-06-23.zip"
 
     mineral_name = table_column.CharField(max_length=75, blank=True, null=True)
     """
@@ -44,7 +50,7 @@ class AflowPrototype(Structure):
     Pearson symbol for the prototype structure
     """
 
-    strukturbericht = table_column.CharField(max_length=6)
+    strukturbericht_symbol = table_column.CharField(max_length=6)
     """
     Strukturbericht symbol for the prototype structure
     """
@@ -70,3 +76,12 @@ class AflowPrototype(Structure):
             return f"{self.mineral_name} ({self.formula_reduced}) Structure-type"
         else:
             return f"{self.formula_reduced} Structure-type"
+
+    @property
+    def external_link(self) -> str:
+        """
+        URL to this prototype in the AFLOW website.
+        """
+        # All COD structures have their data mapped to a URL in the same way
+        # ex: http://www.aflowlib.org/prototype-encyclopedia/A2B_hP9_150_ef_bd.html"
+        return f"http://www.aflowlib.org/prototype-encyclopedia/{self.aflow_id}.html"
