+++ src/simmate/database/third_parties/aflow/structures.py
