--- src/simmate/toolkit/creators/structure/prototypes/ordered_and_disordered.py
@@ -1,1 +1,0 @@
-# -*- coding: utf-8 -*-
