--- src/simmate/website/prototypes/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class PrototypesConfig(AppConfig):
-
-    # use the full import path for this app b/c it's within a package
-    name = "simmate.website.prototypes"
